+++ app/components/freeWatchList/freeWatchList.tsx
@@ -1,0 +1,25 @@
+"use client";
+
+import useFreeWatchList from "./useFreeWatchList";
+
+import HomeTitleBar from "../homeTitleBar";
+import HomeList from "../homeList";
+
+import homeStyles from "@styles/pages/home/home.module.scss";
+
+export default function FreeWatchList({ list }: { list: ContentsDataType[] }) {
+  const { listData, listRef, tabData, currentTab, onClickTab } =
+    useFreeWatchList(list);
+
+  return (
+    <div className={homeStyles.list_container}>
+      <HomeTitleBar
+        title="Free To Watch"
+        tabData={tabData}
+        currentTab={currentTab}
+        onClick={onClickTab}
+      />
+      <HomeList listData={listData} listRef={listRef} />
+    </div>
+  );
+}
