+++ app/components/popularList/usePopularList.ts
@@ -1,14 +1,9 @@
 "use client";
 
 import { useRef, useState } from "react";
-
 import { popularListApi } from "@/api/httpClient";
-import CategoryTab from "./categoryTab";
-import HomeList from "./homeList";
 
-import contentsStyles from "@styles/pages/home/contents.module.scss";
-
-export default function PopularList({ list }: { list: ContentsDataType[] }) {
+export default function usePopularList(list: ContentsDataType[]) {
   const listRef = useRef<HTMLUListElement>(null);
 
   const [listData, setListData] = useState<ContentsDataType[]>(list);
@@ -31,17 +26,11 @@ export default function PopularList({ list }: { list: ContentsDataType[] }) {
     });
   };
 
-  return (
-    <div className={contentsStyles.content}>
-      <div className={contentsStyles.title_bar}>
-        <h2>{`What's Popular`}</h2>
-        <CategoryTab
-          tabData={tabData}
-          currentTab={currentTab}
-          onClick={onClickTab}
-        />
-      </div>
-      <HomeList listData={listData} listRef={listRef} />
-    </div>
-  );
+  return {
+    listRef,
+    listData,
+    currentTab,
+    tabData,
+    onClickTab,
+  };
 }
