--- components/pages/home/categoryTab.tsx
@@ -1,34 +1,0 @@
-import categoryTabStyles from "@styles/pages/home/categoryTab.module.scss";
-
-type TabData = {
-  name: string;
-  type: string;
-};
-
-export default function CategoryTab({
-  tabData,
-  currentTab,
-  onClick,
-}: {
-  tabData: TabData[];
-  currentTab: string;
-  onClick: (type: string) => void;
-}) {
-  return (
-    <div className={categoryTabStyles.category_tab}>
-      {tabData.map((val: TabData) => {
-        const { name, type } = val;
-
-        return (
-          <button
-            key={type}
-            className={currentTab === type ? categoryTabStyles.current_tab : ""}
-            onClick={() => onClick(type)}
-          >
-            <div>{name}</div>
-          </button>
-        );
-      })}
-    </div>
-  );
-}
