--- styles/pages/home/contents.module.scss
@@ -1,13 +1,0 @@
-.content {
-  padding-top: 30px;
-}
-
-.title_bar {
-  display: flex;
-  gap: 20px;
-  padding: 0 40px;
-
-  h2 {
-    margin: 0;
-  }
-}
